# API Marshaler: Add generated marshalers for RESTJSON protocol