--- a/private/protocol/restxml/build_test.go
+++ b/private/protocol/restxml/build_test.go
@@ -4056,7 +4056,6 @@ func (s *InputService19TestShapeGrant) MarshalFields(e protocol.FieldEncoder) er
 	if s.Grantee != nil {
 		v := s.Grantee
 		attrs := make([]protocol.Attribute, 0, 1)
-		// TODO only create array if an attribute value is set
 
 		if s.Grantee.Type != nil {
 			v := *s.Grantee.Type