… private/protocol/json/encode_test.go
… private/protocol/restjson/build_test.go
… private/protocol/restjson/encode_test.go
… private/protocol/restjson/unmarshal_test.go
… service/apigateway/api.go
… service/batch/api.go
… service/clouddirectory/api.go
… service/cloudsearchdomain/api.go
… service/cognitosync/api.go
… service/efs/api.go
… service/elasticsearchservice/api.go
… service/elastictranscoder/api.go
… service/glacier/api.go
… service/greengrass/api.go
… service/iot/api.go
… service/lambda/api.go
… service/lexmodelbuildingservice/api.go
… service/lexruntimeservice/api.go
… service/mobile/api.go
… service/mobileanalytics/api.go
… service/pinpoint/api.go
… service/polly/api.go
… service/workdocs/api.go
… service/xray/api.go