Updates the RESTJSON protocol marshalers to use code generation instead
of reflection based marshaling. These changes improve the performance of
the SDK's marshaling in both time (CPU), number and size of allocations.

The following benchmark shows a relative difference between the SDKs
previous marshaling of RESTJSON performance and the performance with the
generated marshaling.

```
benchmark                                                   old ns/op     new ns/op     delta
BenchmarkRESTJSONBuild_Complex_ETCCreateJob-4               105896        36199         -65.82%
BenchmarkRESTJSONBuild_Simple_ETCListJobsByPipeline-4       8246          5085          -38.33%
BenchmarkRESTJSONRequest_Complex_CFCreateJob-4              318451        215346        -32.38%
BenchmarkRESTJSONRequest_Simple_ETCListJobsByPipeline-4     133614        126604        -5.25%

benchmark                                                   old allocs     new allocs     delta
BenchmarkRESTJSONBuild_Complex_ETCCreateJob-4               327            239            -26.91%
BenchmarkRESTJSONBuild_Simple_ETCListJobsByPipeline-4       67             50             -25.37%
BenchmarkRESTJSONRequest_Complex_CFCreateJob-4              686            598            -12.83%
BenchmarkRESTJSONRequest_Simple_ETCListJobsByPipeline-4     245            228            -6.94%

benchmark                                                   old bytes     new bytes     delta
BenchmarkRESTJSONBuild_Complex_ETCCreateJob-4               24112         19992         -17.09%
BenchmarkRESTJSONBuild_Simple_ETCListJobsByPipeline-4       3728          3765          +0.99%
BenchmarkRESTJSONRequest_Complex_CFCreateJob-4              74537         70378         -5.58%
BenchmarkRESTJSONRequest_Simple_ETCListJobsByPipeline-4     49721         49764         +0.09%
```